Parallel stream proc doesn't inject index now
It was too confusing, it just splits stream into n concurrent streams, if you
need to know original location within the source stream you can wrap it in
enumerate quite easily. The idea is that stream consists of "work items" and
each item has enough information to complete (i.e. contains information for
destination location of the result, which could be an index or some more
complicated structure)